# Improve JSON validation perf