--- a/src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteProperties.Helpers.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteProperties.Helpers.cs
@@ -63,7 +63,7 @@ namespace System.Text.Json
         [MethodImpl(MethodImplOptions.NoInlining)]
         private void OnValidateWritingPropertyFailed()
         {
-            if (_enclosingContainer == EnclosingContainerType.PartialValue)
+            if (IsWritingPartialString)
             {
                 ThrowInvalidOperationException(ExceptionResource.CannotWriteWithinString);
             }