… src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.WriteValues.StringSegment.cs
… src/libraries/System.Text.Json/src/System/Text/Json/Writer/Utf8JsonWriter.cs