Fixes #111039

With the addition of string value segment APIs to the JSON writer (#101356), validation in UTF8JsonWriter needs to also validate that values are not written after uncompleted string segments. This extra check for every write impacts perf, especially when writing values in tight loops. The linked bug above shows some such scenarios (e.g. serializing collections). This PR improves performance for validation by (1) refactoring exception throwing call sites and (2) changing the validation check when writing JSON values.

### Perf

The commits measured are main (bc2ac2f), pre-StringValueSegment (before the regression was introduced, e5878e9), and this PR. All 4 test cases from #111039 are shown here. 

#### WriteJson<HashSet<String>>.SerializeToWriter

| Method            | Toolchain                                                                                                         | Mode      | Mean     | Ratio |
|------------------ |------------------------------------------------------------------------------------------------------------------ |---------- |---------:|------:|
| SerializeToWriter | main | SourceGen | 1.774 us |  1.12  |
| SerializeToWriter | pre-StringValueSegment   | SourceGen | 1.583 us |  1.00 |
| SerializeToWriter | PR       | SourceGen |1.456 us |  0.92 |

#### WriteJson<ImmutableDictionary<String, String>>.SerializeToWriter

| Method            | Toolchain                                                                                                         | Mode      | Mean     | Ratio |
|------------------ |------------------------------------------------------------------------------------------------------------------ |---------- |---------:|------:|
| SerializeToWriter | main | SourceGen | 5.966 us |  1.33 |
| SerializeToWriter | pre-StringValueSegment | SourceGen | 4.483 us |  1.00 |
| SerializeToWriter | PR       | SourceGen | 4.450 us |  0.99 |

#### WriteJson\<LoginViewModel\>.SerializeToWriter

| Method            | Toolchain                                                                                                         | Mode      | Mean     | Ratio |
|------------------ |------------------------------------------------------------------------------------------------------------------ |---------- |---------:|------:|
| SerializeToWriter | main | SourceGen | 64.03 ns |  1.07 |
| SerializeToWriter | pre-StringValueSegment  | SourceGen | 59.82 ns |  1.00 |
| SerializeToWriter | PR       | SourceGen | 56.46 ns |  0.94 |

#### Perf_Basic.WriteBasicUtf8

| Method         | Toolchain                                                                                                         | Formatted | SkipValidation | DataSize | Mean     | Ratio |
|--------------- |------------------------------------------------------------------------------------------------------------------ |---------- |--------------- |--------- |---------:|------:|
| WriteBasicUtf8 | main | False     | False          | 100000   | 984.8 us |  1.01 |
| WriteBasicUtf8 | pre-StringValueSegment  | False     | False          | 100000   |978.0 us |  1.00 |
| WriteBasicUtf8 | PR       | False     | False          | 100000   | 771.0 us |  0.79 |

### Changes in this PR

#### Exceptions

Exceptions are now using the throw helper pattern with minimal arguments. This reduces the code size in the validation methods and defers the performance penalty to only the exceptional scenarios. With this change the JIT also seems more willing to inline business logic which gives perf improvements in tight loop source generated methods like this:
```cs
private void HashSetStringSerializeHandler(global::System.Text.Json.Utf8JsonWriter writer, global::System.Collections.Generic.HashSet<string>? value)
{
    if (value is null)
    {
        writer.WriteNullValue();
        return;
    }

    writer.WriteStartArray();

    foreach (string element in value)
    {
        writer.WriteStringValue(element);
    }

    writer.WriteEndArray();
}
```

#### Validation for values

The validation check for writing values is now a single check instead of multiple checks based on enclosing container type. For this, `_isObject` is now changed to `_enclosingContainer`, which can be either object, array, root/none, or partial value. This removes the need to check `CurrentDepth` separately to determine whether the writer is at the root level. In addition, the actual enum values are:
```cs
/// <summary>
/// The type of container that is enclosing the current position. The underlying values have been chosen
/// to allow <see cref="CanWriteValue"/> to be done using bitwise operations and must be kept in sync with <see cref="JsonTokenType"/>.
/// </summary>
internal enum EnclosingContainerType : byte
{
    /// <summary>
    /// Root level.
    /// </summary>
    None = JsonTokenType.None,

    /// <summary>
    /// JSON object.
    /// </summary>
    Object = JsonTokenType.PropertyName,

    /// <summary>
    /// JSON array. This is chosen large enough to not conflict with any <see cref="JsonTokenType"/> value.
    /// </summary>
    Array =                 0b000_1_0000,

    /// <summary>
    /// Partial UTF-8 string. This is a container if viewed as an array of "utf-8 string segment"-typed values. This array can only be one level deep
    /// so <see cref="_bitStack"/> does not need to store its state.
    /// <see cref="IsWritingPartialString"/> relies on the value of the partial string members being the largest values of this enum.
    /// </summary>
    Utf8StringSequence =    0b001_0_0000,

    /// <summary>
    /// Partial UTF-16 string. This is a container if viewed as an array of "utf-16 string segment"-typed values. This array can only be one level deep
    /// so <see cref="_bitStack"/> does not need to store its state.
    /// <see cref="IsWritingPartialString"/> relies on the value of the partial string members being the largest values of this enum.
    /// </summary>
    Utf16StringSequence =   0b010_0_0000,

    /// <summary>
    /// Partial Base64 string. This is a container if viewed as an array of "base64 string segment"-typed values. This array can only be one level deep
    /// so <see cref="_bitStack"/> does not need to store its state.
    /// <see cref="IsWritingPartialString"/> relies on the value of the partial string members being the largest values of this enum.
    /// </summary>
    Base64StringSequence =  0b011_0_0000,
}
```

These were chosen in such a way that the value validation can be reduced to:

```cs
private bool CanWriteValue => _enclosingContainer == EnclosingContainerType.Array | (byte)_enclosingContainer == (byte)_tokenType;
```

Writing values in arrays is always valid and writing values in partial values is never valid. Otherwise, in objects and at the root level, the only valid previous tokens are property and none respectively. This can be checked with equality since we picked the enum values for `EnclosingContainerType` for this purpose.