@eiriktsarpalis 2025-01-13
> Here is a rough inductive proof that this works:

Nice! Would the hypothetical inclusion of new token types in the future render this optimization obsolete?

Would it be possible to simplify this technique if we moved `StringSegmentSentinel` to instead be represented as enclosing container type(s)? Technically a sequence of string segments does denote an array-like container where you can only append segments of the same encoding.