@PranavSenthilnathan 2025-01-13
> Would it be possible to simplify this technique if we moved `StringSegmentSentinel` to instead be represented as enclosing container type(s)? Technically a sequence of string segments does denote an array-like container where you can only append segments of the same encoding.

Yep, that is much simpler, I updated PR description with the changes and the perf (which didn't change much).

> Nice! Would the hypothetical inclusion of new token types in the future render this optimization obsolete?

The new version still depends on `JsonTokenType.None` and `JsonTokenType.PropertyName` but I don't think it will be difficult to find another simple bit-check if we do add something new. The previous code certainly risked this issue.