@eiriktsarpalis 2025-01-14
I would recommend merging #111041 before continuing work on this PR. Functional completeness first, then we can make it faster.