@PranavSenthilnathan 2025-01-17
> Do you have updated benchmark results to share after including the most recent changes?

Updated results for latest commit. Looks basically the same as before.

> I would also recommend comparing the full benchmark suite from `dotnet/performance` in case a particular workflow has slipped.

Only `Perf_ParseThenWrite` seemed to have a new regression, around 5% slower. After talking offline, we decided to get this PR in and reassess if a regression shows up in the perf CI.